+++ server/routes/profiles.js
@@ -1,6 +1,6 @@
 const app = require('express').Router();
 const bodyParser = require('body-parser');
-const { profiles, hashtagsOnProfiles } = require('../models');
+const { profiles, hashtagsOnProfiles, bookmarksOnPost } = require('../models');
 const { Op } = require('sequelize');
 const { nullCheck, defaultNullFields } = require('./validations/nullcheck');
 const { jwtcheck, authorizeProfileUUID } = require('../middleware/jwtcheck');
@@ -21,6 +21,9 @@ app.get("/:profileUUID", async (req, res) => {
         }
     })
         .then((result) => {
+            if (result == null) {
+                return res.status(409).send("invalid resource");
+            }
             res.send(result);
         })
         .catch((err) => {
@@ -38,7 +41,7 @@ app.post("/", jwtcheck, async (req, res) => {
 
     await profiles.create(req.body)
         .then((result) => {
-            res.send("profile created successfully!!");
+            res.send("SUCCESS");
         })
         .catch((err) => {
             res.status(403).send(err.message);
@@ -53,15 +56,20 @@ app.put("/:profileUUID", jwtcheck, authorizeProfileUUID, async (req, res) => {
     value = nullCheck(req.body, { mustBeNullFields: [...defaultNullFields, 'userId', 'followers', 'followings', 'isActive'] });
     if (typeof (value) == 'string') return res.status(409).send(value);
 
-    const profileUUID = req.params.profileUUID;
+    const profileId = req.userinfo.profileId;
 
     await profiles.update(req.body, {
         where: {
-            "uuid": profileUUID,
+            "id": {
+                [Op.eq]: profileId
+            },
         }
     })
         .then((result) => {
-            res.send("profile updated successfully!!");
+            if (result == 0) {
+                return res.status(409).send("invalid resource");
+            }
+            res.send("SUCCESS");
         })
         .catch((err) => {
             res.status(403).send(err.message);
@@ -73,17 +81,20 @@ app.put("/:profileUUID", jwtcheck, authorizeProfileUUID, async (req, res) => {
 * @check check jwt signature, match profile uuid from payload
 */
 app.delete("/:profileUUID", jwtcheck, authorizeProfileUUID, async (req, res) => {
-    const profileUUID = req.params.profileUUID;
+    const profileId = req.userinfo.profileId;
 
     profiles.destroy({
         where: {
-            "uuid": {
-                [Op.eq]: profileUUID,
+            "id": {
+                [Op.eq]: profileId,
             }
         }
     })
         .then((result) => {
-            res.send("profile deleted successfully!!");
+            if (result == 0) {
+                return res.status(409).send("invalid resource");
+            }
+            res.send("SUCCESS");
         })
         .catch((err) => {
             res.status(403).send(err.message);
@@ -114,6 +125,10 @@ app.get("/:profileUUID/hashtags", async (req, res) => {
 
     if (error) return;
 
+    if (result == null) {
+        return res.status(409).send("invalid resource");
+    }
+
     const profileId = result.id;
 
     await hashtagsOnProfiles.findAll({
@@ -138,26 +153,9 @@ app.get("/:profileUUID/hashtags", async (req, res) => {
 * @check check jwt signature, match profile uuid from payload
 */
 app.post("/:profileUUID/hashtags/:hashtagUUID", jwtcheck, authorizeProfileUUID, async (req, res) => {
-    const profileUUID = req.params.profileUUID;
+    const profileId = req.userinfo.profileId;
     const hashtagUUID = req.params.hashtagUUID;
     let error = false;
-
-    result = await profiles.findOne({
-        where: {
-            "uuid": {
-                [Op.eq]: profileUUID,
-            },
-        },
-        attributes: ['id'],
-    })
-        .catch((err) => {
-            error = true;
-            res.status(403).send(err.message);
-        });
-
-    if (error) return;
-
-    const profileId = result.id;
 
     result = await hashtags.findOne({
         where: {
@@ -174,12 +172,15 @@ app.post("/:profileUUID/hashtags/:hashtagUUID", jwtcheck, authorizeProfileUUID, 
 
     if (error) return;
 
+    if (result == null) {
+        return res.status(409).send("invalid resource");
+    }
 
     const hashtagId = result.id;
 
     await hashtagsOnProfiles.create({ "profileId": profileId, "hashtagId": hashtagId })
         .then((result) => {
-            res.send("hashtag added successfully!!");
+            res.send("SUCCESS");
         })
         .catch((err) => {
             res.status(403).send(err.message);
@@ -191,26 +192,9 @@ app.post("/:profileUUID/hashtags/:hashtagUUID", jwtcheck, authorizeProfileUUID, 
 * @check check jwt signature, match profile uuid from payload
 */
 app.delete("/:profileUUID/hashtags/hashtagUUID", jwtcheck, authorizeProfileUUID, async (req, res) => {
-    const profileUUID = req.params.profileUUID;
+    const profileId = req.userinfo.profileId;
     const hashtagUUID = req.params.hashtagUUID;
     let error = false;
-
-    result = await profiles.findOne({
-        where: {
-            "uuid": {
-                [Op.eq]: profileUUID,
-            },
-        },
-        attributes: ['id'],
-    })
-        .catch((err) => {
-            error = true;
-            res.status(403).send(err.message);
-        });
-
-    if (error) return;
-
-    const profileId = result.id;
 
     result = await hashtags.findOne({
         where: {
@@ -226,6 +210,10 @@ app.delete("/:profileUUID/hashtags/hashtagUUID", jwtcheck, authorizeProfileUUID,
         });
 
     if (error) return;
+
+    if (result == null) {
+        return res.status(409).send("invalid resource");
+    }
 
     await hashtagsOnProfiles.destroy({
         where: {
@@ -238,10 +226,34 @@ app.delete("/:profileUUID/hashtags/hashtagUUID", jwtcheck, authorizeProfileUUID,
         }
     })
         .then((result) => {
-            res.send("hashtag removed successfully!!");
+            if (result == 0) {
+                return res.status(409).send("invalid resource");
+            }
+            res.send("SUCCESS");
         })
         .catch((err) => {
             res.status(403).send(err.message);
+        });
+});
+
+/* 
+* /:profileUUID/bookmarks/posts - GET - get all bookmarked posts of user
+* @check check jwt signature, match profileuuid with payload
+*/
+app.get("/:profileUUID/bookmarks/posts", jwtcheck, authorizeProfileUUID, async (req, res) => {
+    const profileId = req.userinfo.profileId;
+
+    await bookmarksOnPost.findAll({
+        where: {
+            "profileId": profileId,
+        },
+        attributes: ['id'],
+    })
+        .then((result) => {
+            res.send(result);
+        })
+        .catch((err) => {
+            res.send(err);
         });
 });
